@@ -4,6 +4,7 @@
   "version": "1.0.0",
   "description": "stainless api monorepo",
   "author": "dev@stainlessapi.com",
+  "packageManager": "pnpm@8.6.7",
   "repository": {
     "type": "git",
     "url": "https://github.com/stainless-api/stl-api.git"
@@ -16,9 +17,19 @@
     "stainless",
     "api"
   ],
+  "config": {
+    "ghooks": {
+      "pre-commit": "lint-staged"
+    }
+  },
+  "lint-staged": {
+    "*.{js,mjs,cjs,ts,mts,cts,tsx,json,css,md}": [
+      "prettier --write"
+    ]
+  },
   "license": "ISC",
   "scripts": {
-    "test": "jest",
+    "test": "pnpm -r test",
     "test:ci": "start-server-and-test start-test-server http://0.0.0.0:3005 test",
     "start-test-server": "cd demo && pnpm start -- -p 3005",
     "build": "pnpm -r run prebuild && tsc -b packages/*/tsconfig*.json && (cd demo && pnpm build)",
@@ -28,21 +39,21 @@
     "docs:deploy": "pnpm docs:build && netlify deploy --dir docs/build --prod",
     "watch": "pnpm run build:watch",
     "clean": "pnpm -r run clean && rm -rf packages/*/{package,*.tgz}",
-    "format": "prettier --write ."
+    "format": "prettier --write .",
+    "prettier:check": "prettier --list-different .",
+    "postinstall": "./node_modules/ghooks/bin/install"
   },
   "devDependencies": {
-    "@types/jest": "^29.5.1",
     "@types/node": "^20.2.3",
     "@types/node-fetch": "^2.6.4",
     "execa": "^7.1.1",
-    "jest": "^29.5.0",
-    "jest-watch-typeahead": "^2.2.2",
+    "ghooks": "^2.0.4",
+    "lint-staged": "^13.2.3",
     "lodash": "^4.17.21",
     "node-fetch": "^2.6.11",
     "prettier": "^2.8.8",
     "rimraf": "^5.0.1",
     "start-server-and-test": "^2.0.0",
-    "ts-jest": "^29.1.0",
-    "typescript": "^5.0.4"
+    "typescript": "^5.1.3"
   }
 }
