@@ -27,7 +27,7 @@
   "devDependencies": {
     "@types/lodash": "^4.14.194",
     "@types/node": "^20.2.3",
-    "typescript": "^5.0.4"
+    "typescript": "^5.1.3"
   },
   "dependencies": {
     "stainless": "workspace:*",
