@@ -38,17 +38,27 @@
     }
   },
   "scripts": {
-    "prebuild": "mkdir -p dist/esm && echo '{\"type\": \"module\"}' > dist/esm/package.json"
+    "clean": "rimraf dist *.tsbuildinfo",
+    "prebuild": "mkdir -p dist/esm && echo '{\"type\": \"module\"}' > dist/esm/package.json",
+    "test": "jest"
   },
   "devDependencies": {
     "@tanstack/query-core": "^4.29.5",
     "@tanstack/react-query": "^4.29.5",
+    "@testing-library/react": "^14.0.0",
+    "@types/jest": "^29.5.1",
     "@types/lodash": "^4.14.194",
     "@types/node": "^20.2.3",
     "@types/qs": "^6.9.7",
     "@types/react": "18.0.28",
+    "@types/react-test-renderer": "^18.0.0",
+    "jest": "^29.5.0",
+    "jest-environment-jsdom": "^29.6.1",
+    "jest-watch-typeahead": "^2.2.2",
     "react": "^18.2.0",
-    "typescript": "^5.0.4"
+    "react-test-renderer": "^18.2.0",
+    "ts-jest": "^29.1.0",
+    "typescript": "^5.1.3"
   },
   "dependencies": {
     "lodash": "^4.17.21",
