@@ -18,21 +18,38 @@
   ],
   "license": "ISC",
   "source": "src/stl.ts",
+  "bin": {
+    "gen-stl-api-route-map": "./dist/cli/gen-stl-api-route-map.js"
+  },
   "main": "dist/stl.js",
   "types": "dist/stl.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/stl.d.ts",
+      "default": "./dist/stl.js"
+    }
+  },
   "scripts": {
     "clean": "rimraf dist *.tsbuildinfo",
-    "prebuild": "mkdir -p dist/esm && echo '{\"type\": \"module\"}' > dist/esm/package.json"
+    "prebuild": "mkdir -p dist/esm && echo '{\"type\": \"module\"}' > dist/esm/package.json",
+    "test": "jest"
   },
   "devDependencies": {
+    "@types/jest": "^29.5.1",
     "@types/lodash": "^4.14.194",
     "@types/node": "^20.2.3",
     "@types/qs": "^6.9.7",
-    "typescript": "^5.0.4"
+    "jest": "^29.5.0",
+    "jest-environment-jsdom": "^29.6.1",
+    "jest-watch-typeahead": "^2.2.2",
+    "ts-jest": "^29.1.0",
+    "typescript": "^5.1.3"
   },
   "dependencies": {
+    "dedent-js": "^1.0.1",
     "lodash": "^4.17.21",
     "qs": "^6.11.2",
+    "ts-node": "^10.9.1",
     "zod": "^3.21.4",
     "zod-openapi": "github:stainless-api/zod-openapi#2.5.1"
   }
