@@ -1,7 +1,8 @@
 import useSWR from "swr";
-import { client } from "../api/client";
+import { useClient } from "../api/client";
 
 const useUsers = () => {
+  const client = useClient();
   const { data, error, isLoading, mutate } = useSWR("users-list", () =>
     client.users.list({ pageSize: 3 })
   );
