@@ -2,6 +2,7 @@ import { isEmpty } from "lodash";
 import {
   AnyEndpoint,
   AnyResourceConfig,
+  EndpointHasRequiredQuery,
   EndpointQueryInput,
   EndpointResponseOutput,
   z,
@@ -30,16 +31,27 @@ export type ClientUseInfiniteQuery<
   TQueryData = TQueryFnData
 > = E["path"] extends z.ZodTypeAny
   ? E["query"] extends z.ZodTypeAny
-    ? (
-        path: EndpointPathParam<E>,
-        query: EndpointQueryInput<E>,
-        options?: UseInfiniteQueryOptions<
-          TQueryFnData,
-          TError,
-          TData,
-          TQueryData
-        >
-      ) => UseInfiniteQueryResult<TData, TError>
+    ? EndpointHasRequiredQuery<E> extends true
+      ? (
+          path: EndpointPathParam<E>,
+          query: EndpointQueryInput<E>,
+          options?: UseInfiniteQueryOptions<
+            TQueryFnData,
+            TError,
+            TData,
+            TQueryData
+          >
+        ) => UseInfiniteQueryResult<TData, TError>
+      : (
+          path: EndpointPathParam<E>,
+          query?: EndpointQueryInput<E>,
+          options?: UseInfiniteQueryOptions<
+            TQueryFnData,
+            TError,
+            TData,
+            TQueryData
+          >
+        ) => UseInfiniteQueryResult<TData, TError>
     : (
         path: EndpointPathParam<E>,
         options?: UseInfiniteQueryOptions<
@@ -50,10 +62,25 @@ export type ClientUseInfiniteQuery<
         >
       ) => UseInfiniteQueryResult<TData, TError>
   : E["query"] extends z.ZodTypeAny
-  ? (
-      query: EndpointQueryInput<E>,
-      options?: UseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryData>
-    ) => UseInfiniteQueryResult<TData, TError>
+  ? EndpointHasRequiredQuery<E> extends true
+    ? (
+        query: EndpointQueryInput<E>,
+        options?: UseInfiniteQueryOptions<
+          TQueryFnData,
+          TError,
+          TData,
+          TQueryData
+        >
+      ) => UseInfiniteQueryResult<TData, TError>
+    : (
+        query?: EndpointQueryInput<E>,
+        options?: UseInfiniteQueryOptions<
+          TQueryFnData,
+          TError,
+          TData,
+          TQueryData
+        >
+      ) => UseInfiniteQueryResult<TData, TError>
   : (
       options?: UseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryData>
     ) => UseInfiniteQueryResult<TData, TError>;
