@@ -4,6 +4,7 @@ import {
   type EndpointBodyInput,
   type EndpointQueryInput,
   type EndpointResponseOutput,
+  EndpointHasRequiredQuery,
 } from "stainless";
 import { isEmpty } from "lodash";
 import {
@@ -19,58 +20,78 @@ export type ClientUseQuery<
   TData = TQueryFnData
 > = E["path"] extends z.ZodTypeAny
   ? E["body"] extends z.ZodTypeAny
-    ? (
-        path: EndpointPathParam<E>,
-        body: EndpointBodyInput<E>,
-        options?: { query?: EndpointQueryInput<E> } & UseQueryOptions<
-          TQueryFnData,
-          TError,
-          TData
-        >
-      ) => UseQueryResult<TData, TError>
+    ? EndpointHasRequiredQuery<E> extends true
+      ? (
+          path: EndpointPathParam<E>,
+          body: EndpointBodyInput<E>,
+          options: UseQueryOptions<E, TQueryFnData, TError, TData>
+        ) => UseQueryResult<TData, TError>
+      : (
+          path: EndpointPathParam<E>,
+          body: EndpointBodyInput<E>,
+          options?: UseQueryOptions<E, TQueryFnData, TError, TData>
+        ) => UseQueryResult<TData, TError>
     : E["query"] extends z.ZodTypeAny
-    ? (
-        path: EndpointPathParam<E>,
-        query: EndpointQueryInput<E>,
-        options?: UseQueryOptions<TQueryFnData, TError, TData>
-      ) => UseQueryResult<TData, TError>
+    ? EndpointHasRequiredQuery<E> extends true
+      ? (
+          path: EndpointPathParam<E>,
+          query: EndpointQueryInput<E>,
+          options?: BaseUseQueryOptions<TQueryFnData, TError, TData>
+        ) => UseQueryResult<TData, TError>
+      : (
+          path: EndpointPathParam<E>,
+          query?: EndpointQueryInput<E>,
+          options?: BaseUseQueryOptions<TQueryFnData, TError, TData>
+        ) => UseQueryResult<TData, TError>
     : (
         path: EndpointPathParam<E>,
-        options?: UseQueryOptions<TQueryFnData, TError, TData>
+        options?: UseQueryOptions<E, TQueryFnData, TError, TData>
       ) => UseQueryResult<TData, TError>
   : E["body"] extends z.ZodTypeAny
-  ? (
-      body: EndpointBodyInput<E>,
-      options?: { query?: EndpointQueryInput<E> } & UseQueryOptions<
-        TQueryFnData,
-        TError,
-        TData
-      >
-    ) => UseQueryResult<TData, TError>
+  ? EndpointHasRequiredQuery<E> extends true
+    ? (
+        body: EndpointBodyInput<E>,
+        options: UseQueryOptions<E, TQueryFnData, TError, TData>
+      ) => UseQueryResult<TData, TError>
+    : (
+        body: EndpointBodyInput<E>,
+        options?: UseQueryOptions<E, TQueryFnData, TError, TData>
+      ) => UseQueryResult<TData, TError>
   : E["query"] extends z.ZodTypeAny
-  ? (
-      query: EndpointQueryInput<E>,
-      options?: UseQueryOptions<TQueryFnData, TError, TData>
-    ) => UseQueryResult<TData, TError>
+  ? EndpointHasRequiredQuery<E> extends true
+    ? (
+        query: EndpointQueryInput<E>,
+        options?: BaseUseQueryOptions<TQueryFnData, TError, TData>
+      ) => UseQueryResult<TData, TError>
+    : (
+        query?: EndpointQueryInput<E>,
+        options?: BaseUseQueryOptions<TQueryFnData, TError, TData>
+      ) => UseQueryResult<TData, TError>
   : (
-      options?: UseQueryOptions<TQueryFnData, TError, TData>
+      options?: UseQueryOptions<E, TQueryFnData, TError, TData>
     ) => UseQueryResult<TData, TError>;
 
 export type UseQueryOptions<
-  TQueryFnData = unknown,
+  E extends AnyEndpoint,
+  TQueryFnData = EndpointResponseOutput<E>,
   TError = unknown,
   TData = TQueryFnData
 > = Omit<
   BaseUseQueryOptions<TQueryFnData, TError, TData, any>,
   "queryKey" | "queryFn"
->;
+> &
+  (E["query"] extends z.ZodTypeAny
+    ? EndpointHasRequiredQuery<E> extends true
+      ? { query: EndpointQueryInput<E> }
+      : { query?: EndpointQueryInput<E> }
+    : {});
 
 export type KeysEnum<T> = { [P in keyof Required<T>]: true };
 
 // This is required so that we can determine if a given object matches the RequestOptions
 // type at runtime. While this requires duplication, it is enforced by the TypeScript
 // compiler such that any missing / extraneous keys will cause an error.
-const useQueryOptionsKeys: KeysEnum<{ query?: any } & UseQueryOptions> = {
+const useQueryOptionsKeys: KeysEnum<{ query?: any } & UseQueryOptions<any>> = {
   context: true,
   retry: true,
   retryDelay: true,
