@@ -35,12 +35,12 @@ it(`generating in folder`, async () =>
 {
   "stl-api-gen/src/__tests__/common.ts": "import { z } from "zod";
 import { Enum as __enum_Enum } from "../../../src/__tests__/common";
-export const AddressIface = z.object({ street: z.string(), city: z.string(), state: z.string(), postalCode: z.string() });
-export const Enum = z.nativeEnum(__enum_Enum);
+export const AddressIface: z.ZodTypeAny = z.object({ street: z.string(), city: z.string(), state: z.string(), postalCode: z.string() });
+export const Enum: z.ZodTypeAny = z.nativeEnum(__enum_Enum);
 ",
   "stl-api-gen/src/__tests__/generate-in-folder.test.ts": "import { z } from "zod";
 import { AddressIface as __symbol_AddrIface, Enum as __enum_Enum } from "./common";
-const T = z.object({ datetime: z.string().max(20).datetime().regex(new RegExp("[A-Z]"), "oops"), addresses: z.array(z.lazy(() => __symbol_AddrIface)), enum: z.lazy(() => __enum_Enum) });
+const T: z.ZodTypeAny = z.object({ datetime: z.string().max(20).datetime().regex(new RegExp("[A-Z]"), "oops"), addresses: z.array(z.lazy(() => __symbol_AddrIface)), enum: z.lazy(() => __enum_Enum) });
 ",
 }
 `
