@@ -18,32 +18,17 @@ interface T {
 
 it(`imported interface`, async () =>
   expect(
-  await multiFileTestCase({
-    __filename
-  })
-).toMatchInlineSnapshot(`
+    await multiFileTestCase({
+      __filename,
+    })
+  ).toMatchInlineSnapshot(`
 {
-  "src/__tests__/common.codegen.d.ts": "import { z } from "zod";
-export const AddressIface: z.ZodTypeAny;
+  "src/__tests__/common.codegen.ts": "import { z } from "zod";
+export const AddressIface: z.ZodTypeAny = z.object({ street: z.string(), city: z.string(), state: z.string(), postalCode: z.string() });
 ",
-  "src/__tests__/common.codegen.js": "const { z } = require("zod");
-const AddressIface = z.object({ street: z.string(), city: z.string(), state: z.string(), postalCode: z.string() });
-exports.AddressIface = AddressIface;
-",
-  "src/__tests__/common.codegen.mjs": "import { z } from "zod";
-export const AddressIface = z.object({ street: z.string(), city: z.string(), state: z.string(), postalCode: z.string() });
-",
-  "src/__tests__/imported-aliased-interface.test.codegen.d.ts": "import { z } from "zod";
+  "src/__tests__/imported-aliased-interface.test.codegen.ts": "import { z } from "zod";
 import { AddressIface as __symbol_Addr } from "./common";
-const T: z.ZodTypeAny;
-",
-  "src/__tests__/imported-aliased-interface.test.codegen.js": "const { z } = require("zod");
-const { AddressIface: __symbol_Addr } = require("./common");
-const T = z.object({ firstName: z.string(), lastName: z.string(), address: z.lazy(() => __symbol_Addr).optional() });
-",
-  "src/__tests__/imported-aliased-interface.test.codegen.mjs": "import { z } from "zod";
-import { AddressIface as __symbol_Addr } from "./common";
-const T = z.object({ firstName: z.string(), lastName: z.string(), address: z.lazy(() => __symbol_Addr).optional() });
+const T: z.ZodTypeAny = z.object({ firstName: z.string(), lastName: z.string(), address: z.lazy(() => __symbol_Addr).optional() });
 ",
 }
 `));
